Add a layout visualization tool
Use `./draw.py <keyboard>` to visualize the connections
of the keyboard layout you specified

Additionally, Dvorak is now part of a list of layouts
that have their blank keys on the left rather than
a single special case.  This change also correctly
orients the left and right keys in reversed layouts.
Although the left and right being swapped did not cause
any tests to fail (and therefore did not affect the action
of the cipher), it was important to fix the swap in case
this module is used elsewhere and to be pedantically correct.